Step 6: Add a profile link
Step 6: Add a profile link
add links to your awesome portfolio projects. Links help create context when you’re communicating in issues and pull requests.
You might link to a website, a repository, or even a line of code. To create a link, put the text you want to display in square brackets, and the URL in the following parentheses.

[GitHub](http://github.com) Activity: Add a link
## Edit the file _includes/03-links.md.
Replace the filler text with a link to your GitHub profile (or anywhere else).